+++ ChangeLog.md
@@ -3,6 +3,7 @@
 -------------------------
 ## [Unrelease]
 ### Added
+- Implement method `myDataset::load()` to load mnist dataset.
 
 ### Changed